+++ tests/Rule/General/DisableCallMethodInObjectMethodRule/Source/BaseClassWithCall.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Efabrica\PHPStanRules\Tests\Rule\General\DisableCallMethodInObjectMethodRule\Source;
+
+abstract class BaseClassWithCall implements WithCallInterface
+{
+}