+++ tests/Rule/General/DisableCallMethodInObjectMethodRule/Source/UncheckedClass.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Efabrica\PHPStanRules\Tests\Rule\General\DisableCallMethodInObjectMethodRule\Source;
+
+class UncheckedClass
+{
+    public function disabledMethod(): UncheckedClass
+    {
+        return $this;
+    }
+}