+++ tests/Rule/General/DisableCallMethodInObjectMethodRule/Source/WithDisabledMethodInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Efabrica\PHPStanRules\Tests\Rule\General\DisableCallMethodInObjectMethodRule\Source;
+
+interface WithDisabledMethodInterface
+{
+
+}